Multilanguage component added
